requirements.txt
apache-airflow==2.9.1
pandas>=2.2.0
pyarrow>=16.0.0
requests>=2.32.0
urllib3>=2.2.0

# Testing
pytest>=8.2.0
pytest-cov>=5.0.0
